Finish edit

+++ src/App.js
@@ -15,11 +15,19 @@ function App() {
   const [isAddingMovie, setIsAddingMovie] = useState(false);
 
   const onDeleteMovies = () => {
+    const markedForDeletion = movies.filter((m) => m.markedForDeletion);
+
+    if (!markedForDeletion.length) {
+      alert("Select movies first!");
+
+      return;
+    }
+
     const result = window.confirm("Are you sure you want to delete selected movies?");
 
     if (!result) return;
 
-    const promises = movies.filter((m) => m.markedForDeletion).map(m => {
+    const promises = markedForDeletion.map(m => {
       return fetch(`${process.env.REACT_APP_API_HOST}/movies/${m.id}`, {
         method: "DELETE",
         headers: {
